chore(weak): move to `weak` directory.
