+++ src/lib/weak/index.ts
@@ -1,0 +1,4 @@
+export { IndexedNamedWeakData  } from './indexed-named-weak-data.class';
+export { IndexedWeakData } from './indexed-weak-data.class';
+export { NamedWeakData } from './named-weak-data.class';
+export { WeakData } from './weak-data.class';
